trading_bot/requirements.txt
# Telegram Bot
pyTelegramBotAPI==4.14.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Database
mysql-connector-python==8.2.0

# Data Analysis & Technical Indicators
pandas==2.1.4
numpy==1.24.3
pandas-ta==0.3.14b0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
xgboost==2.0.2

# Plotting & Charts
matplotlib==3.8.2
mplfinance==0.12.10b0

# News API
newsapi-python==0.2.6

# Scheduling
schedule==1.2.0

# Data Classes & Type Hints
typing-extensions==4.8.0

# Date/Time utilities
python-dateutil==2.8.2